# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mealprep API
  version: 0.2.0
  description: |
    Energy is kcal, mass is grams, volume is millilitres. A product's macros and
    nutrients are per 100 of its unit; a serving amount is in that same unit.
    Every entity carries createdAt and updatedAt; deletes are soft and deleted
    entities are never returned.
servers:
  - url: /v1
...
paths:
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List the user's products
      responses:
        "200":
          description: The products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListProductsResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product by hand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: The created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /products/search:
    get:
      operationId: searchProducts
      tags: [products]
      summary: Search the Open Food Facts catalog
      parameters:
        - name: q
          in: query
          description: Free text matched against catalog product names
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The catalog entries matching the query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchProductsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"
  /products/import:
    post:
      operationId: importProduct
      tags: [products]
      summary: Snapshot a catalog entry into the user's products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportProductRequest"
      responses:
        "201":
          description: The imported product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: The catalog entry lacks one of the four macros
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          $ref: "#/components/responses/BadGateway"
  /products/{productId}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      operationId: getProduct
      tags: [products]
      summary: Fetch one product
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Replace a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateProductResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Soft-delete a product
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    ProductId:
      name: productId
      in: path
      description: The product
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: No such entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Open Food Facts could not be reached or answered with an error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Unit:
      type: string
      enum: [g, ml]
    Macros:
      type: object
      required: [energyKcal, fatG, proteinG, carbohydratesG]
      properties:
        energyKcal:
          type: number
          format: double
        fatG:
          type: number
          format: double
        proteinG:
          type: number
          format: double
        carbohydratesG:
          type: number
          format: double
    Nutrients:
      type: object
      description: Every further nutrient per 100 units, keyed by name
      additionalProperties:
        type: number
        format: double
    Product:
      type: object
      required: [id, name, unit, macros, nutrients, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        unit:
          $ref: "#/components/schemas/Unit"
        macros:
          $ref: "#/components/schemas/Macros"
        nutrients:
          $ref: "#/components/schemas/Nutrients"
        brand:
          type: string
          description: Who makes it; snapshotted at import and part of the product's identity, not its presentation
        sourceCode:
          type: string
          description: Open Food Facts code the snapshot came from; absent when created by hand
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductDraft:
      type: object
      required: [name, unit, macros]
      properties:
        name:
          type: string
        unit:
          $ref: "#/components/schemas/Unit"
        macros:
          $ref: "#/components/schemas/Macros"
        nutrients:
          $ref: "#/components/schemas/Nutrients"
        brand:
          type: string
    CatalogEntry:
      type: object
      required: [code, name, unit, macros, nutrients, complete]
      properties:
        code:
          type: string
        name:
          type: string
        unit:
          $ref: "#/components/schemas/Unit"
        macros:
          $ref: "#/components/schemas/Macros"
        nutrients:
          $ref: "#/components/schemas/Nutrients"
        brand:
          type: string
          description: First of the catalog's brand tags; several entries often share a name and differ only by this
        imageUrl:
          type: string
          format: uri
          description: Catalog-hosted thumbnail, linked rather than copied because the images carry a different licence from the data
        complete:
          type: boolean
          description: False when one of the four macros is missing, which makes the entry unimportable
    ListProductsResponse:
      type: object
      required: [products]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
    GetProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/Product"
    CreateProductRequest:
      $ref: "#/components/schemas/ProductDraft"
    CreateProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/Product"
    UpdateProductRequest:
      $ref: "#/components/schemas/ProductDraft"
    UpdateProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/Product"
    SearchProductsResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CatalogEntry"
    ImportProductRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
    ImportProductResponse:
      type: object
      required: [product]
      properties:
        product:
          $ref: "#/components/schemas/Product"
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string