# backend/internal/repository/sqlite/db/models.gen.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql"
)

type CalendarDay struct {
	Date      string
	DayPlanID string
	CreatedAt string
	UpdatedAt string
	DeletedAt sql.NullString
}

type DayPlan struct {
	ID        string
	Label     string
	CreatedAt string
	UpdatedAt string
	DeletedAt sql.NullString
}

type DayPlanMeal struct {
	ID        string
	DayPlanID string
	MealID    string
	Position  int64
	CreatedAt string
	UpdatedAt string
	DeletedAt sql.NullString
}

type Meal struct {
	ID        string
	Label     string
	CreatedAt string
	UpdatedAt string
	DeletedAt sql.NullString
}

type MealServing struct {
	ID        string
	MealID    string
	ProductID string
	Amount    float64
	Position  int64
	CreatedAt string
	UpdatedAt string
	DeletedAt sql.NullString
}

type Product struct {
	ID             string
	Name           string
	Unit           string
	EnergyKcal     float64
	FatG           float64
	ProteinG       float64
	CarbohydratesG float64
	Nutrients      string
	Brand          sql.NullString
	SourceCode     sql.NullString
	CreatedAt      string
	UpdatedAt      string
	DeletedAt      sql.NullString
}

type Target struct {
	ID             int64
	EnergyKcal     float64
	FatG           float64
	ProteinG       float64
	CarbohydratesG float64
	CreatedAt      string
	UpdatedAt      string
	DeletedAt      sql.NullString
}